Slide "matchinfo" outward when loading
This change makes the "matchinfo" field from the "Bracket" tab in the dashboard software behave similarly to the updated "VS" part in the Away scene. 

TLDR, if "matchinfo" is found it will swipe outward from its center and won't be visible until the animation occurs (previously, it would be permanently visible on screen before anything else appeared).